update hình ảnh sản phẩm và hiệu ứng ajax

+++ Controllers/HangHoaController.cs
@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using WebsiteE_Commerce.Data;
@@ -12,95 +8,71 @@ namespace WebsiteE_Commerce.Controllers
     public class HangHoaController : Controller
     {
         private readonly QuanLyHangHoaContext _context;
+        private readonly IWebHostEnvironment _host;
 
-        public HangHoaController(QuanLyHangHoaContext context)
+        public HangHoaController(QuanLyHangHoaContext context, IWebHostEnvironment host)
         {
             _context = context;
+            _host = host;
         }
 
-        // GET: HangHoa
+        // Hiện danh sách hàng hóa
         public async Task<IActionResult> Index()
         {
-            var quanLyHangHoaContext = _context.HangHoas.Include(h => h.MaLoaiNavigation).Include(h => h.MaNccNavigation);
-            return View(await quanLyHangHoaContext.ToListAsync());
+            var list = await _context.HangHoas.Include(h => h.MaLoaiNavigation).ToListAsync();
+            return View(list);
         }
 
-        // GET: HangHoa/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var hangHoa = await _context.HangHoas
                 .Include(h => h.MaLoaiNavigation)
                 .Include(h => h.MaNccNavigation)
                 .FirstOrDefaultAsync(m => m.MaHh == id);
-            if (hangHoa == null)
-            {
-                return NotFound();
-            }
-
+            if (hangHoa == null) return NotFound();
             return View(hangHoa);
         }
 
-        // GET: HangHoa/Create
+        [HttpGet]
         public IActionResult Create()
         {
-            ViewData["MaLoai"] = new SelectList(_context.Loais, "MaLoai", "MaLoai");
-            ViewData["MaNcc"] = new SelectList(_context.NhaCungCaps, "MaNcc", "MaNcc");
+            ViewBag.MaLoai = new SelectList(_context.Loais, "MaLoai", "TenLoai");
             return View();
         }
 
-        // POST: HangHoa/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("MaHh,TenHh,TenAlias,MaLoai,MoTaDonVi,DonGia,Hinh,NgaySx,GiamGia,SoLanXem,MoTa,MaNcc")] HangHoa hangHoa)
+        public async Task<IActionResult> Create(HangHoa hangHoa)
         {
+            ModelState.Remove("MaLoaiNavigation");
+            ModelState.Remove("MaNccNavigation");
             if (ModelState.IsValid)
             {
                 _context.Add(hangHoa);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["MaLoai"] = new SelectList(_context.Loais, "MaLoai", "MaLoai", hangHoa.MaLoai);
-            ViewData["MaNcc"] = new SelectList(_context.NhaCungCaps, "MaNcc", "MaNcc", hangHoa.MaNcc);
+            ViewBag.MaLoai = new SelectList(_context.Loais, "MaLoai", "TenLoai", hangHoa.MaLoai);
             return View(hangHoa);
         }
 
-        // GET: HangHoa/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var hangHoa = await _context.HangHoas.FindAsync(id);
-            if (hangHoa == null)
-            {
-                return NotFound();
-            }
-            ViewData["MaLoai"] = new SelectList(_context.Loais, "MaLoai", "MaLoai", hangHoa.MaLoai);
-            ViewData["MaNcc"] = new SelectList(_context.NhaCungCaps, "MaNcc", "MaNcc", hangHoa.MaNcc);
+            if (hangHoa == null) return NotFound();
+            ViewBag.MaLoai = new SelectList(_context.Loais, "MaLoai", "TenLoai", hangHoa.MaLoai);
             return View(hangHoa);
         }
 
-        // POST: HangHoa/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("MaHh,TenHh,TenAlias,MaLoai,MoTaDonVi,DonGia,Hinh,NgaySx,GiamGia,SoLanXem,MoTa,MaNcc")] HangHoa hangHoa)
+        public async Task<IActionResult> Edit(int id, HangHoa hangHoa)
         {
-            if (id != hangHoa.MaHh)
-            {
-                return NotFound();
-            }
-
+            if (id != hangHoa.MaHh) return NotFound();
+            ModelState.Remove("MaLoaiNavigation");
+            ModelState.Remove("MaNccNavigation");
             if (ModelState.IsValid)
             {
                 try
@@ -110,60 +82,60 @@ namespace WebsiteE_Commerce.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!HangHoaExists(hangHoa.MaHh))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!_context.HangHoas.Any(e => e.MaHh == hangHoa.MaHh)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["MaLoai"] = new SelectList(_context.Loais, "MaLoai", "MaLoai", hangHoa.MaLoai);
-            ViewData["MaNcc"] = new SelectList(_context.NhaCungCaps, "MaNcc", "MaNcc", hangHoa.MaNcc);
+            ViewBag.MaLoai = new SelectList(_context.Loais, "MaLoai", "TenLoai", hangHoa.MaLoai);
             return View(hangHoa);
         }
 
-        // GET: HangHoa/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var hangHoa = await _context.HangHoas
-                .Include(h => h.MaLoaiNavigation)
-                .Include(h => h.MaNccNavigation)
-                .FirstOrDefaultAsync(m => m.MaHh == id);
-            if (hangHoa == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var hangHoa = await _context.HangHoas.Include(h => h.MaLoaiNavigation).FirstOrDefaultAsync(m => m.MaHh == id);
+            if (hangHoa == null) return NotFound();
             return View(hangHoa);
         }
 
-        // POST: HangHoa/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var hangHoa = await _context.HangHoas.FindAsync(id);
-            if (hangHoa != null)
-            {
-                _context.HangHoas.Remove(hangHoa);
-            }
-
+            if (hangHoa != null) _context.HangHoas.Remove(hangHoa);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
-        private bool HangHoaExists(int id)
+        // --- AJAX: Kiểm tra trùng tên ---
+        [HttpGet]
+        public IActionResult KiemTraTenAjax(string tenSp)
         {
-            return _context.HangHoas.Any(e => e.MaHh == id);
+            var exists = _context.HangHoas.Any(x => x.TenHh == tenSp);
+            return Json(new { isExisted = exists });
+        }
+
+        // --- AJAX: Upload ảnh ---
+        [HttpPost]
+        public IActionResult UploadHinhAjax(IFormFile fHinh)
+        {
+            if (fHinh != null)
+            {
+                string fileName = Guid.NewGuid().ToString() + Path.GetExtension(fHinh.FileName);
+                string path = Path.Combine(_host.WebRootPath, "images", fileName);
+                if (!Directory.Exists(Path.Combine(_host.WebRootPath, "images")))
+                {
+                    Directory.CreateDirectory(Path.Combine(_host.WebRootPath, "images"));
+                }
+                using (var stream = new FileStream(path, FileMode.Create))
+                {
+                    fHinh.CopyTo(stream);
+                }
+                return Json(new { success = true, fileName = fileName });
+            }
+            return Json(new { success = false });
         }
     }
 }
